test(chart): recategorize and clean up browser tests
