+++ packages/react/src/components/chart/radar-chart.test.tsx
@@ -1,4 +1,4 @@
-import { page, render } from "#test/browser"
+import { render, screen } from "#test"
 import { RadarChart } from "."
 
 interface Data {
@@ -16,59 +16,50 @@ const data: Data[] = [
 const RADAR_DOM_NODES_PER_SERIES = 2
 
 describe("<RadarChart />", () => {
-  test("sets `displayName` correctly", () => {
-    expect(RadarChart.Root.displayName).toBe("RadarChart")
-  })
-
-  test("renders generated radars and default grid/angle axis", async () => {
+  test("renders generated radars and default grid/angle axis", () => {
     const series: Array<{ dataKey: "downloads" | "visits" }> = [
       { dataKey: "visits" },
       { dataKey: "downloads" },
     ]
 
-    await render(
+    render(
       <RadarChart.Root
         data-testid="root"
         data={data}
         nameKey="browser"
         series={series}
+        responsiveContainerProps={{ children: null, height: 400, width: 400 }}
       />,
     )
 
-    const root = page.getByTestId("root")
+    const root = screen.getByTestId("root")
 
-    await expect.element(root).toHaveClass("ui-radar-chart")
-    await expect
-      .poll(
-        () => root.element().querySelectorAll(".ui-polar-chart__radar").length,
-      )
-      .toBe(series.length * RADAR_DOM_NODES_PER_SERIES)
-    await expect
-      .poll(
-        () => root.element().querySelectorAll(".ui-polar-chart__grid").length,
-      )
-      .toBeGreaterThan(0)
-    await expect
-      .poll(
-        () =>
-          root.element().querySelectorAll(".ui-polar-chart__angle-axis").length,
-      )
-      .toBeGreaterThan(0)
+    expect(root).toHaveClass("ui-radar-chart")
+    expect(root.querySelectorAll(".ui-polar-chart__radar")).toHaveLength(
+      series.length * RADAR_DOM_NODES_PER_SERIES,
+    )
+    expect(
+      root.querySelectorAll(".ui-polar-chart__grid").length,
+    ).toBeGreaterThan(0)
+    expect(
+      root.querySelectorAll(".ui-polar-chart__angle-axis").length,
+    ).toBeGreaterThan(0)
   })
 
-  test("renders composition components instead of fallback `series`", async () => {
+  test("renders composition components instead of fallback `series`", () => {
     const fallbackSeries: Array<{ dataKey: "downloads" | "visits" }> = [
       { dataKey: "visits" },
       { dataKey: "downloads" },
     ]
     const composedSeries: Array<{ dataKey: "visits" }> = [{ dataKey: "visits" }]
 
-    await render(
+    render(
       <RadarChart.Root
         data-testid="root"
         data={data}
         nameKey="browser"
         series={fallbackSeries}
+        responsiveContainerProps={{ children: null, height: 400, width: 400 }}
       >
         {composedSeries.map((series) => (
           <RadarChart.Radar key={series.dataKey} dataKey={series.dataKey} />
@@ -76,12 +67,10 @@ describe("<RadarChart />", () => {
       </RadarChart.Root>,
     )
 
-    const root = page.getByTestId("root")
+    const root = screen.getByTestId("root")
 
-    await expect
-      .poll(
-        () => root.element().querySelectorAll(".ui-polar-chart__radar").length,
-      )
-      .toBe(composedSeries.length * RADAR_DOM_NODES_PER_SERIES)
+    expect(root.querySelectorAll(".ui-polar-chart__radar")).toHaveLength(
+      composedSeries.length * RADAR_DOM_NODES_PER_SERIES,
+    )
   })
 })
